# internal/api/docs/swagger.yaml
definitions:
...
  api.GetDailyFeedOutput:
    properties:
      daily_feed:
        $ref: '#/definitions/db.DailyFeed'
    type: object
  api.GetDailyFeedsOutput:
    properties:
      daily_feeds:
        items:
          $ref: '#/definitions/db.DailyFeed'
        type: array
    type: object
  api.GetExpenseOutput:
    properties:
      expense:
        $ref: '#/definitions/db.Expense'
    type: object
  api.GetExpensesOutput:
    properties:
      expenses:
        items:
          $ref: '#/definitions/db.Expense'
        type: array
    type: object
...
  api.UpsertDailyFeedInput:
    properties:
      animalid:
        type: string
      feed_amount:
        type: number
      feed_date:
        type: string
      feedid:
        type: string
      feedpurchaseid:
        type: string
      projectid:
        type: string
    required:
    - animalid
    - feed_amount
    - feed_date
    - feedid
    - feedpurchaseid
    - projectid
    type: object
  api.UpsertExpenseInput:
    properties:
      cost:
        type: number
      date:
        type: string
      items:
        type: string
      projectid:
        type: string
      quantity:
        type: number
    required:
    - cost
    - date
    - items
    - projectid
    - quantity
    type: object
...
  db.DailyFeed:
    properties:
      animalid:
        type: string
      created:
        type: string
      feed_amount:
        type: number
      feed_date:
        type: string
      feedid:
        type: string
      feedpurchaseid:
        type: string
      id:
        type: string
      projectid:
        type: string
      updated:
        type: string
      userid:
        type: string
    type: object
  db.Expense:
    properties:
      cost:
        type: number
      created:
        type: string
      date:
        type: string
      id:
        type: string
      items:
        type: string
      projectid:
        type: string
      quantity:
        type: number
      updated:
        type: string
      userid:
        type: string
    type: object
...
info:
  contact: {}
  title: 4H Record Books API
paths:
...
  /daily-feed:
    get:
      consumes:
      - application/json
      description: Gets all of a user's daily feeds for a given project and animal
      parameters:
      - description: Project ID
        in: query
        name: projectId
        required: true
        type: string
      - description: Animal ID
        in: query
        name: animalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetDailyFeedsOutput'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get daily feeds by project and animal
      tags:
      - Daily Feed
    post:
      consumes:
      - application/json
      description: Adds a daily feed to a user's personal records
      parameters:
      - description: Daily Feed information
        in: body
        name: UpsertDailyFeedInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertDailyFeedInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Add a daily feed
      tags:
      - Daily Feed
  /daily-feed/{dailyFeedId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's daily feed given the daily feed ID
      parameters:
      - description: Daily Feed ID
        in: path
        name: dailyFeedId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes a daily feed
      tags:
      - Daily Feed
    get:
      consumes:
      - application/json
      description: Get a user's daily feed by ID
      parameters:
      - description: Daily Feed ID
        in: path
        name: dailyFeedId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetDailyFeedOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get a daily feed
      tags:
      - Daily Feed
    put:
      consumes:
      - application/json
      description: Updates a user's daily feed information
      parameters:
      - description: Daily Feed ID
        in: path
        name: dailyFeedId
        required: true
        type: string
      - description: DailyFeed information
        in: body
        name: UpsertDailyFeedInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertDailyFeedInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update a daily feed
      tags:
      - Daily Feed
  /expense:
    get:
      consumes:
      - application/json
      description: Gets all of a user's expenses given a project ID
      parameters:
      - description: Project ID
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetExpensesOutput'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Get expenses by project
      tags:
      - Expense
    post:
      consumes:
      - application/json
      description: Adds an expense to a user's personal records
      parameters:
      - description: Expense information
        in: body
        name: UpsertExpenseInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertExpenseInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Adds an expense
      tags:
      - Expense
  /expense/{expenseId}:
    delete:
      consumes:
      - application/json
      description: Deletes a user's expense given the expense ID
      parameters:
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Removes an expense
      tags:
      - Expense
    get:
      consumes:
      - application/json
      description: Get a user's expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetExpenseOutput'
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get an expense
      tags:
      - Expense
    put:
      consumes:
      - application/json
      description: Updates a user's expense information
      parameters:
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      - description: Expense information
        in: body
        name: UpsertExpenseInput
        required: true
        schema:
          $ref: '#/definitions/api.UpsertExpenseInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Update an expense
      tags:
      - Expense
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"